Remastered version: modernized the Solidity source
Updated to modern Solidity coding style for this "remastered" 2024
version. This is a squashed commit of 11 commits.

Added version information to README

Listing versions affecting the building process (and the resulting
bytecode).

Remastered version

This newest version is the "remastered" version of the original 2016
code. The remastered version is compliant with modern Solidity coding
conventions, is a Truffle project, and is released under Apache
License 2.0.

Changed line length to 74

Added more comments

Polished the language in the comments

Used ChatGPT o4-mini-high for improving the language.

Formatted lines to ~74 characters

Treating 74 as a soft-limit, as even with diff tags and email quotation
lines seem to fit in 80 characters.

Clarified a comment

Clarified the commenting on resign(), and made the style more
consistent.

README: Updating and clarifying

Clarifying the description of Coinia Vy, and updated the year.

Used npx prettier with the Solidity plug-in

Formatted the Solidity source according to the style guide.

Updated the compiler artifact

Updated the resulting JSON.